refactor(concepts): consolidate concept requirement tests
