+++ clang/test/SemaTemplate/concepts.cpp
@@ -1713,3 +1713,16 @@ template<C<void, bool> T> int f();
 void main() { f<int>(); }
 
 }
+
+namespace GH176402 {
+  void f() {
+    auto recursiveLambda = [](auto self, int depth) -> void {
+      struct MyClass;
+      auto testConcept = []<typename T> {
+        return requires(T) { &MyClass::operator0 }
+        // expected-error {{expected ';' at end of requirement}}
+      };
+    };
+    recursiveLambda(recursiveLambda, 5);
+  }
+}
